no_extract array part1 and fix source check up

+++ misc/scripts/install-local.sh
@@ -1160,10 +1160,12 @@ function hashcheck_down() {
 function genextr_down() {
     genextr_declare
     hashcheck_down
-    fancy_message sub "Extracting ${CYAN}${dest}${NC}"
-    ${ext_method} "${dest}" 1>&1 2> /dev/null
-    if [[ -f ${dest} ]]; then
-        rm -f "${dest}"
+    if ! [[ ${no_extract[*]} == *${dest}* ]]; then
+        fancy_message sub "Extracting ${CYAN}${dest}${NC}"
+        ${ext_method} "${dest}" 1>&1 2> /dev/null
+        if [[ -f ${dest} ]]; then
+            rm -f "${dest}"
+        fi
     fi
     if [[ -z ${source[1]} ]]; then
         cd ./*/ 2> /dev/null || {
